pyproject.toml
[project]
name = "basketball-football-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.4.5",
    "bs4>=0.0.2",
    "duckduckgo-search>=8.0.1",
    "fuzzywuzzy>=0.18.0",
    "groq>=0.24.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "python-levenshtein>=0.27.1",
    "requests==2.27.1",
]
